+++ src/main/javascript/geojs/v1/geoEnrich.js
@@ -10,25 +10,25 @@ License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF 
 See the License for the specific language governing permissions and limitations under the License. */
 
 /**
- * A geoEnhance object
+ * A geoEnrich object
  * @namespace
  */
-GEOAPIS_V1.geoEnhance = function(accessToken){
+GEOAPIS_V1.geoEnrich = function(accessToken){
 	GEOAPIS_V1.baseService.call(this, accessToken);
 };
 /**
- * geoEnhance inherits baseService
+ * geoEnrich inherits baseService
  * @namespace
  */
-GEOAPIS_V1_INHERIT(GEOAPIS_V1.baseService, GEOAPIS_V1.geoEnhance);
+GEOAPIS_V1_INHERIT(GEOAPIS_V1.baseService, GEOAPIS_V1.geoEnrich);
 /**
  * Set API URL to get address
  * @param Object {latitude: float(required), longitude: float(required), searchRadius: number(optional), searchRadiusUnit: string(optional)}
  * @callback {callback: User defined callback} (optional) or
  * @return response object
  */
-GEOAPIS_V1.geoEnhance.prototype.getAddress = function(params, callback){
-	var apiUrl = '/geoenhance/v1/address/bylocation?latitude='+encodeURIComponent(params.latitude)+'&longitude='+encodeURIComponent(params.longitude),
+GEOAPIS_V1.geoEnrich.prototype.getAddress = function(params, callback){
+	var apiUrl = '/geoenrich/v1/address/bylocation?latitude='+encodeURIComponent(params.latitude)+'&longitude='+encodeURIComponent(params.longitude),
 	optionalList = ['searchRadius', 'searchRadiusUnit'];
 	apiUrl = this.apiGetUrl(params, optionalList, apiUrl);
 	if(callback !== undefined){
@@ -45,8 +45,8 @@ GEOAPIS_V1.geoEnhance.prototype.getAddress = function(params, callback){
  * @callback {callback: User defined callback} (optional) or
  * @return response object
  */
-GEOAPIS_V1.geoEnhance.prototype.getPlace = function(params, callback){
-	var apiUrl = '/geoenhance/v1/place/bylocation?latitude='+encodeURIComponent(params.latitude)+'&longitude='+encodeURIComponent(params.longitude),
+GEOAPIS_V1.geoEnrich.prototype.getPlace = function(params, callback){
+	var apiUrl = '/geoenrich/v1/place/bylocation?latitude='+encodeURIComponent(params.latitude)+'&longitude='+encodeURIComponent(params.longitude),
 	optionalList = ['levelHint'];
 	apiUrl = this.apiGetUrl(params, optionalList, apiUrl);
 	if(callback !== undefined){
@@ -63,26 +63,10 @@ GEOAPIS_V1.geoEnhance.prototype.getPlace = function(params, callback){
  * @callback {callback: User defined callback} (optional) or
  * @return response object
  */
-GEOAPIS_V1.geoEnhance.prototype.getPOI = function(params, callback){
-	var apiUrl = '/geoenhance/v1/poi/bylocation?latitude='+encodeURIComponent(params.latitude)+'&longitude='+encodeURIComponent(params.longitude),
+GEOAPIS_V1.geoEnrich.prototype.getPOI = function(params, callback){
+	var apiUrl = '/geoenrich/v1/poi/bylocation?latitude='+encodeURIComponent(params.latitude)+'&longitude='+encodeURIComponent(params.longitude),
 	optionalList = ['category', 'searchRadius', 'maxCandidates', 'searchRadiusUnit', 'searchPriority', 'searchDataset', 'brandName', 'travelTime', 'travelTimeUnit', 'travelDistance', 'travelDistanceUnit', 'mode'];
 	apiUrl = this.apiGetUrl(params, optionalList, apiUrl);
-	if(callback !== undefined){
-		this.callApiAsync(apiUrl, callback);
-	}
-	else{
-		this.callApi(apiUrl);
-		return this.response;
-	}
-};
-/**
- * Set API URL to get timezone
- * @param Object {latitude: float(required), longitude: float(required), timestamp: float(required)}
- * @callback {callback: User defined callback} (optional) or
- * @return response object
- */
-GEOAPIS_V1.geoEnhance.prototype.getTimezone = function(params, callback){
-	var apiUrl = '/geoenhance/v1/timezone/bylocation?latitude='+encodeURIComponent(params.latitude)+'&longitude='+encodeURIComponent(params.longitude)+'&timestamp='+encodeURIComponent(params.timestamp);
 	if(callback !== undefined){
 		this.callApiAsync(apiUrl, callback);
 	}