+++ CHANGELOG.md
@@ -25,4 +25,4 @@
 ## 2.0.0 - 2016-07-04
 
 ### Added
-- Added 2.0 Version of JavaScript SDK which has Location Intelligence API features GeoEnhance,GeoLife,GeoSearch,Geo911,GeoTax and Geocode.
+- Added 2.0 Version of Java SDK which has Location Intelligence API features GeoEnhance,GeoLife,GeoSearch,Geo911,GeoTax and Geocode.